clean up unitstr usage, closes #292
