+++ gpkit/varkey.py
@@ -1,6 +1,7 @@
 """Defines the VarKey class"""
 from .small_classes import Strings, Quantity, Count
-from .small_scripts import unitstr, veckeyed
+from .small_scripts import veckeyed
+from .repr_conventions import unitstr
 
 
 class VarKey(object):
@@ -96,13 +97,11 @@ class VarKey(object):
     def __getattr__(self, attr):
         return self.descr.get(attr, None)
 
-    def unitstr(self, dimless=""):
-        "Returns string representation of units"
-        return unitstr(self, into=" [%s] ", dimless=dimless)
+    unitstr = unitstr
 
     def latex_unitstr(self):
         "Returns latex unitstr"
-        us = unitstr(self.units, r"~\mathrm{%s}", "L~")
+        us = self.unitstr(r"~\mathrm{%s}", "L~")
         utf = us.replace("frac", "tfrac").replace(r"\cdot", r"\cdot ")
         return utf if utf != r"~\mathrm{-}" else ""
